Section-21: Part-265 -> Submitting the payment

+++ skinet/client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -87,11 +87,22 @@ export class CheckoutPaymentComponent implements OnInit {
 
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: (order) => {
-        this.toastr.show('Order created successfully');
-        console.log(order);
-        this.basketService.deleteLocalBasket();
-        const navigationExtras: NavigationExtras = { state: order };
-        this.router.navigate(['checkout/success'], navigationExtras);
+        this.toastr.success('Order created successfully');
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            }
+          }
+        }).then(result => {
+          console.log(result);
+          if(result.paymentIntent){
+            this.basketService.deleteLocalBasket();
+            const navigationExtras: NavigationExtras = { state: order };
+            this.router.navigate(['checkout/success'], navigationExtras);
+          }
+        });
       },
     });
   }
